requirements.txt
networkx
pyaml
pyvis
rdflib
